small repair

+++ Services/Response/Sale/UserSalesResponse.cs
@@ -7,7 +7,7 @@
         public required string Nip { get; init; }
         public required string Status { get; init; }
         public required DateTime CloseDate { get; init; }
-        public required decimal Value { get; init; }
+        public required long Value { get; init; }
         public required int DecimalPlace { get; init; }
         public required string Currency { get; init; }
         public required string CompanyName { get; init; }
